Split TagLib metadata import/export code into multiple files
