+++ src/sources/metadatasourcetaglib.cpp
@@ -1,6 +1,6 @@
 #include "sources/metadatasourcetaglib.h"
 
-#include "track/trackmetadatataglib.h"
+#include "track/taglib/trackmetadata.h"
 
 #include "util/logger.h"
 #include "util/memory.h"
@@ -119,21 +119,21 @@ MetadataSourceTagLib::importTrackMetadataAndCoverImage(
     switch (m_fileType) {
     case taglib::FileType::MP3: {
         TagLib::MPEG::File file(TAGLIB_FILENAME_FROM_QSTRING(m_fileName));
-        if (!taglib::readAudioProperties(pTrackMetadata, file)) {
+        if (!taglib::readAudioPropertiesFromFile(pTrackMetadata, file)) {
             break;
         }
         // ID3v2 tag takes precedence over APE tag
         if (taglib::hasID3v2Tag(file)) {
             const TagLib::ID3v2::Tag* pTag = file.ID3v2Tag();
             DEBUG_ASSERT(pTag);
-            taglib::importTrackMetadataFromID3v2Tag(pTrackMetadata, *pTag);
-            taglib::importCoverImageFromID3v2Tag(pCoverImage, *pTag);
+            taglib::id3v2::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            taglib::id3v2::importCoverImageFromTag(pCoverImage, *pTag);
             return afterImport(ImportResult::Succeeded);
         } else if (taglib::hasAPETag(file)) {
             const TagLib::APE::Tag* pTag = file.APETag();
             DEBUG_ASSERT(pTag);
-            taglib::importTrackMetadataFromAPETag(pTrackMetadata, *pTag);
-            taglib::importCoverImageFromAPETag(pCoverImage, *pTag);
+            taglib::ape::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            taglib::ape::importCoverImageFromTag(pCoverImage, *pTag);
             return afterImport(ImportResult::Succeeded);
         } else if (taglib::hasID3v1Tag(file)) {
             // Note (TagLib 1.1.11): TagLib::MPEG::File::tag() returns a
@@ -150,21 +150,21 @@ MetadataSourceTagLib::importTrackMetadataAndCoverImage(
     }
     case taglib::FileType::MP4: {
         TagLib::MP4::File file(TAGLIB_FILENAME_FROM_QSTRING(m_fileName));
-        if (!taglib::readAudioProperties(pTrackMetadata, file)) {
+        if (!taglib::readAudioPropertiesFromFile(pTrackMetadata, file)) {
             break;
         }
         if (taglib::hasMP4Tag(file)) {
             const TagLib::MP4::Tag* pTag = file.tag();
             DEBUG_ASSERT(pTag);
-            taglib::importTrackMetadataFromMP4Tag(pTrackMetadata, *pTag);
-            taglib::importCoverImageFromMP4Tag(pCoverImage, *pTag);
+            taglib::mp4::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            taglib::mp4::importCoverImageFromTag(pCoverImage, *pTag);
             return afterImport(ImportResult::Succeeded);
         }
         break;
     }
     case taglib::FileType::FLAC: {
         TagLib::FLAC::File file(TAGLIB_FILENAME_FROM_QSTRING(m_fileName));
-        if (!taglib::readAudioProperties(pTrackMetadata, file)) {
+        if (!taglib::readAudioPropertiesFromFile(pTrackMetadata, file)) {
             break;
         }
         bool importSucceeded = false;
@@ -173,14 +173,14 @@ MetadataSourceTagLib::importTrackMetadataAndCoverImage(
         if (taglib::hasXiphComment(file)) {
             TagLib::Ogg::XiphComment* pTag = file.xiphComment();
             DEBUG_ASSERT(pTag);
-            taglib::importTrackMetadataFromVorbisCommentTag(pTrackMetadata, *pTag);
-            coverImageImported = taglib::importCoverImageFromVorbisCommentTag(pCoverImage, *pTag);
+            taglib::xiph::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            coverImageImported = taglib::xiph::importCoverImageFromTag(pCoverImage, *pTag);
             importSucceeded = true;
         } else if (taglib::hasID3v2Tag(file)) {
             const TagLib::ID3v2::Tag* pTag = file.ID3v2Tag();
             DEBUG_ASSERT(pTag);
-            taglib::importTrackMetadataFromID3v2Tag(pTrackMetadata, *pTag);
-            coverImageImported = taglib::importCoverImageFromID3v2Tag(pCoverImage, *pTag);
+            taglib::id3v2::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            coverImageImported = taglib::id3v2::importCoverImageFromTag(pCoverImage, *pTag);
             importSucceeded = true;
         }
         // Only import cover images from picture list as a fallback if file tags
@@ -192,7 +192,7 @@ MetadataSourceTagLib::importTrackMetadataAndCoverImage(
                 importSucceeded &&
                 !coverImageImported) { // no cover image found in file tags
             // Read cover art directly from the file as a fallback
-            *pCoverImage = taglib::importCoverImageFromVorbisCommentPictureList(file.pictureList());
+            *pCoverImage = taglib::xiph::importCoverImageFromPictureList(file.pictureList());
         }
         if (importSucceeded) {
             return afterImport(ImportResult::Succeeded);
@@ -201,13 +201,13 @@ MetadataSourceTagLib::importTrackMetadataAndCoverImage(
     }
     case taglib::FileType::OGG: {
         TagLib::Ogg::Vorbis::File file(TAGLIB_FILENAME_FROM_QSTRING(m_fileName));
-        if (!taglib::readAudioProperties(pTrackMetadata, file)) {
+        if (!taglib::readAudioPropertiesFromFile(pTrackMetadata, file)) {
             break;
         }
         TagLib::Ogg::XiphComment* pTag = file.tag();
         if (pTag) {
-            taglib::importTrackMetadataFromVorbisCommentTag(pTrackMetadata, *pTag);
-            taglib::importCoverImageFromVorbisCommentTag(pCoverImage, *pTag);
+            taglib::xiph::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            taglib::xiph::importCoverImageFromTag(pCoverImage, *pTag);
             return afterImport(ImportResult::Succeeded);
         }
         break;
@@ -215,13 +215,13 @@ MetadataSourceTagLib::importTrackMetadataAndCoverImage(
 #if (TAGLIB_HAS_OPUSFILE)
     case taglib::FileType::OPUS: {
         TagLib::Ogg::Opus::File file(TAGLIB_FILENAME_FROM_QSTRING(m_fileName));
-        if (!taglib::readAudioProperties(pTrackMetadata, file)) {
+        if (!taglib::readAudioPropertiesFromFile(pTrackMetadata, file)) {
             break;
         }
         TagLib::Ogg::XiphComment* pTag = file.tag();
         if (pTag) {
-            taglib::importTrackMetadataFromVorbisCommentTag(pTrackMetadata, *pTag);
-            taglib::importCoverImageFromVorbisCommentTag(pCoverImage, *pTag);
+            taglib::xiph::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            taglib::xiph::importCoverImageFromTag(pCoverImage, *pTag);
             return afterImport(ImportResult::Succeeded);
         }
         break;
@@ -229,21 +229,21 @@ MetadataSourceTagLib::importTrackMetadataAndCoverImage(
 #endif // TAGLIB_HAS_OPUSFILE
     case taglib::FileType::WV: {
         TagLib::WavPack::File file(TAGLIB_FILENAME_FROM_QSTRING(m_fileName));
-        if (!taglib::readAudioProperties(pTrackMetadata, file)) {
+        if (!taglib::readAudioPropertiesFromFile(pTrackMetadata, file)) {
             break;
         }
         if (taglib::hasAPETag(file)) {
             const TagLib::APE::Tag* pTag = file.APETag();
             DEBUG_ASSERT(pTag);
-            taglib::importTrackMetadataFromAPETag(pTrackMetadata, *pTag);
-            taglib::importCoverImageFromAPETag(pCoverImage, *pTag);
+            taglib::ape::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            taglib::ape::importCoverImageFromTag(pCoverImage, *pTag);
             return afterImport(ImportResult::Succeeded);
         }
         break;
     }
     case taglib::FileType::WAV: {
         TagLib::RIFF::WAV::File file(TAGLIB_FILENAME_FROM_QSTRING(m_fileName));
-        if (!taglib::readAudioProperties(pTrackMetadata, file)) {
+        if (!taglib::readAudioPropertiesFromFile(pTrackMetadata, file)) {
             break;
         }
         if (taglib::hasID3v2Tag(file)) {
@@ -253,27 +253,27 @@ MetadataSourceTagLib::importTrackMetadataAndCoverImage(
             const TagLib::ID3v2::Tag* pTag = file.tag();
 #endif
             DEBUG_ASSERT(pTag);
-            taglib::importTrackMetadataFromID3v2Tag(pTrackMetadata, *pTag);
-            taglib::importCoverImageFromID3v2Tag(pCoverImage, *pTag);
+            taglib::id3v2::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            taglib::id3v2::importCoverImageFromTag(pCoverImage, *pTag);
             return afterImport(ImportResult::Succeeded);
         } else if (file.hasInfoTag()) {
             const TagLib::RIFF::Info::Tag* pTag = file.InfoTag();
             DEBUG_ASSERT(pTag);
-            taglib::importTrackMetadataFromRIFFTag(pTrackMetadata, *pTag);
+            taglib::riff::importTrackMetadataFromTag(pTrackMetadata, *pTag);
             return afterImport(ImportResult::Succeeded);
         }
         break;
     }
     case taglib::FileType::AIFF: {
         AiffFile file(TAGLIB_FILENAME_FROM_QSTRING(m_fileName));
-        if (!taglib::readAudioProperties(pTrackMetadata, file)) {
+        if (!taglib::readAudioPropertiesFromFile(pTrackMetadata, file)) {
             break;
         }
         if (taglib::hasID3v2Tag(file)) {
             const TagLib::ID3v2::Tag* pTag = file.tag();
             DEBUG_ASSERT(pTag);
-            taglib::importTrackMetadataFromID3v2Tag(pTrackMetadata, *pTag);
-            taglib::importCoverImageFromID3v2Tag(pCoverImage, *pTag);
+            taglib::id3v2::importTrackMetadataFromTag(pTrackMetadata, *pTag);
+            taglib::id3v2::importCoverImageFromTag(pCoverImage, *pTag);
             return afterImport(ImportResult::Succeeded);
         } else if (file.importTrackMetadataFromTextChunks(pTrackMetadata)) {
             return afterImport(ImportResult::Succeeded);
@@ -336,7 +336,7 @@ class MpegTagSaver : public TagSaver {
         if (pFile->isOpen()) {
             TagLib::ID3v2::Tag* pID3v2Tag = nullptr;
             if (taglib::hasAPETag(*pFile)) {
-                if (taglib::exportTrackMetadataIntoAPETag(pFile->APETag(), trackMetadata)) {
+                if (taglib::ape::exportTrackMetadataIntoTag(pFile->APETag(), trackMetadata)) {
                     modifiedTagsBitmask |= TagLib::MPEG::File::APE;
                 }
                 // Only write ID3v2 tag if it already exists
@@ -345,7 +345,7 @@ class MpegTagSaver : public TagSaver {
                 // Get or create ID3v2 tag
                 pID3v2Tag = pFile->ID3v2Tag(true);
             }
-            if (taglib::exportTrackMetadataIntoID3v2Tag(pID3v2Tag, trackMetadata)) {
+            if (taglib::id3v2::exportTrackMetadataIntoTag(pID3v2Tag, trackMetadata)) {
                 modifiedTagsBitmask |= TagLib::MPEG::File::ID3v2;
             }
         }
@@ -375,7 +375,7 @@ class Mp4TagSaver : public TagSaver {
 
   private:
     static bool exportTrackMetadata(TagLib::MP4::File* pFile, const TrackMetadata& trackMetadata) {
-        return pFile->isOpen() && taglib::exportTrackMetadataIntoMP4Tag(pFile->tag(), trackMetadata);
+        return pFile->isOpen() && taglib::mp4::exportTrackMetadataIntoTag(pFile->tag(), trackMetadata);
     }
 
     TagLib::MP4::File m_file;
@@ -405,14 +405,14 @@ class FlacTagSaver : public TagSaver {
         if (pFile->isOpen()) {
             TagLib::Ogg::XiphComment* pXiphComment = nullptr;
             if (taglib::hasID3v2Tag(*pFile)) {
-                modifiedTags |= taglib::exportTrackMetadataIntoID3v2Tag(pFile->ID3v2Tag(), trackMetadata);
+                modifiedTags |= taglib::id3v2::exportTrackMetadataIntoTag(pFile->ID3v2Tag(), trackMetadata);
                 // Only write VorbisComment tag if it already exists
                 pXiphComment = pFile->xiphComment(false);
             } else {
                 // Get or create VorbisComment tag
                 pXiphComment = pFile->xiphComment(true);
             }
-            modifiedTags |= taglib::exportTrackMetadataIntoXiphComment(pXiphComment, trackMetadata);
+            modifiedTags |= taglib::xiph::exportTrackMetadataIntoTag(pXiphComment, trackMetadata);
         }
         return modifiedTags;
     }
@@ -452,7 +452,7 @@ class OggTagSaver : public TagSaver {
         return false;
 #else
         return pFile->isOpen() &&
-                taglib::exportTrackMetadataIntoXiphComment(pFile->tag(), trackMetadata);
+                taglib::xiph::exportTrackMetadataIntoTag(pFile->tag(), trackMetadata);
 #endif
     }
 
@@ -480,7 +480,7 @@ class OpusTagSaver : public TagSaver {
 
   private:
     static bool exportTrackMetadata(TagLib::Ogg::Opus::File* pFile, const TrackMetadata& trackMetadata) {
-        return pFile->isOpen() && taglib::exportTrackMetadataIntoXiphComment(pFile->tag(), trackMetadata);
+        return pFile->isOpen() && taglib::xiph::exportTrackMetadataIntoTag(pFile->tag(), trackMetadata);
     }
 
     TagLib::Ogg::Opus::File m_file;
@@ -507,7 +507,7 @@ class WavPackTagSaver : public TagSaver {
 
   private:
     static bool exportTrackMetadata(TagLib::WavPack::File* pFile, const TrackMetadata& trackMetadata) {
-        return pFile->isOpen() && taglib::exportTrackMetadataIntoAPETag(pFile->APETag(true), trackMetadata);
+        return pFile->isOpen() && taglib::ape::exportTrackMetadataIntoTag(pFile->APETag(true), trackMetadata);
     }
 
     TagLib::WavPack::File m_file;
@@ -547,9 +547,9 @@ class WavTagSaver : public TagSaver {
         if (pFile->isOpen()) {
             // Write into all available tags
 #if (TAGLIB_HAS_WAV_ID3V2TAG)
-            modifiedTags |= taglib::exportTrackMetadataIntoID3v2Tag(pFile->ID3v2Tag(), trackMetadata);
+            modifiedTags |= taglib::id3v2::exportTrackMetadataIntoTag(pFile->ID3v2Tag(), trackMetadata);
 #else
-            modifiedTags |= taglib::exportTrackMetadataIntoID3v2Tag(pFile->tag(), trackMetadata);
+            modifiedTags |= taglib::id3v2::exportTrackMetadataIntoTag(pFile->tag(), trackMetadata);
 #endif
             modifiedTags |= exportTrackMetadataIntoRIFFTag(pFile->InfoTag(), trackMetadata);
         }
@@ -579,7 +579,7 @@ class AiffTagSaver : public TagSaver {
 
   private:
     static bool exportTrackMetadata(TagLib::RIFF::AIFF::File* pFile, const TrackMetadata& trackMetadata) {
-        return pFile->isOpen() && taglib::exportTrackMetadataIntoID3v2Tag(pFile->tag(), trackMetadata);
+        return pFile->isOpen() && taglib::id3v2::exportTrackMetadataIntoTag(pFile->tag(), trackMetadata);
     }
 
     TagLib::RIFF::AIFF::File m_file;
