+++ src/track/taglib/trackmetadata.h
@@ -1,0 +1,13 @@
+#pragma once
+
+///////////////////////////////////////////////////////////////////////
+// The common source for all tag mappings is MusicBrainz Picard:
+// https://picard.musicbrainz.org/docs/mappings/
+///////////////////////////////////////////////////////////////////////
+
+#include "track/taglib/trackmetadata_ape.h"
+#include "track/taglib/trackmetadata_file.h"
+#include "track/taglib/trackmetadata_id3v2.h"
+#include "track/taglib/trackmetadata_mp4.h"
+#include "track/taglib/trackmetadata_riff.h"
+#include "track/taglib/trackmetadata_xiph.h"
