+++ src/track/taglib/trackmetadata_xiph.h
@@ -1,0 +1,32 @@
+#pragma once
+
+#include <taglib/xiphcomment.h>
+
+#include "track/taglib/trackmetadata_common.h"
+
+namespace mixxx {
+
+namespace taglib {
+
+namespace xiph {
+
+bool importCoverImageFromTag(
+        QImage* pCoverArt,
+        TagLib::Ogg::XiphComment& tag);
+
+QImage importCoverImageFromPictureList(
+        const TagLib::List<TagLib::FLAC::Picture*>& pictures);
+
+void importTrackMetadataFromTag(
+        TrackMetadata* pTrackMetadata,
+        const TagLib::Ogg::XiphComment& tag);
+
+bool exportTrackMetadataIntoTag(
+        TagLib::Ogg::XiphComment* pTag,
+        const TrackMetadata& trackMetadata);
+
+} // namespace xiph
+
+} // namespace taglib
+
+} // namespace mixxx
